refactor much of the code and add many comments
